requirements.txt
# Core dependencies
redis==5.0.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.13.1

# Web framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0

# Data processing
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2

# Utilities
python-dotenv==1.0.0
click==8.1.7
python-dateutil==2.8.2
pytz==2023.3

# Development and testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# Security
cryptography==41.0.8
bcrypt==4.1.2

# Optional: for advanced features
celery==5.3.4
flower==2.0.1

